16 спинт

+++ app/src/main/java/orc/zdertis420/playlistmaker/domain/implementations/interactor/PlayerInteractorImplementation.kt
@@ -26,8 +26,7 @@ class PlayerInteractorImplementation(
         return playerRepository.isPlaying()
     }
 
-    override fun getCurrentPosition(): Int {
+    override fun getCurrentPosition(): Long {
         return playerRepository.getCurrentPosition()
     }
-
 }
